requirements.txt
pylint
django
pylint-django
imutils
Pillow
imutils
skimage
opencv-python
opencv-contrib-python
numpy